Backend/agents/data_validation_agent/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openpyxl
duckdb
scikit-learn
python-multipart
python-dateutil
pandera
rapidfuzz
pytest
PyYAML
